Fix MSL gate regressions and file-size limit from contract enforcement
Post-merge MSL gate triage against a fresh origin/main baseline showed
the contract-enforcement commit regressed 7 models (DAE 476 -> 469):

- TYPE-023 'no other elements under a transitively non-replaceable
  constraint' rejected idiomatic MSL sibling redeclares that add public
  members (RobotR3 GearType1 for GearType2, Batteries CellRCStack for
  CellStack). The clause is removed and TYPE-023 re-classified Deferred;
  every major tool accepts these under plug-compatibility. The TYPE-003
  extra-input rule now correctly keys on a declaration equation
  (binding) instead of the auto-populated start attribute, which had
  made it unreachable.
- EQN-028 (ER124) treated size()/ndims() arguments as variability
  carriers, rejecting LossyGear's 'for i in 1:size(...)' inside
  'if ideal' (LossyGearDemo1/2/3, HeatLosses, ActuatorWithNoise).
  Evaluability now walks expressions and treats the array argument of
  size()/ndims() as structural (MLS 3.7.2).

Also splits rumoca-phase-typecheck constant collection into its own
module: lib.rs was 2034 lines, over the 2000-line review-scan hard
limit that fails the Lint & Format job.
